Move common divs to GuideDivs, Reskin BuffUptimeBar's graph
