+++ src/interface/guide/components/BuffUptimeBar.tsx
@@ -4,6 +4,18 @@ import Spell from 'common/SPELLS/Spell';
 import { Tooltip } from 'interface';
 import { formatPercentage } from 'common/format';
 import { TrackedBuffEvent } from 'parser/core/Entity';
+import {
+  SectionContainer,
+  SectionHeader,
+  TitleColumn,
+  SectionTitle,
+  Label,
+  StatsRow,
+  StatCard,
+  StatValue,
+  StatLabel,
+  InsetContainer,
+} from './GuideDivs';
 
 // Unified uptime graph - handles both simple buffs and stacked buffs
 function UptimeGraph({
@@ -188,13 +200,13 @@ export default function BuffUptimeBar({
   const defaultTooltip = `This is the average number of stacks you had over the course of the fight, counting periods where you didn't have the buff as zero stacks.`;
 
   return (
-    <Container>
-      <TopSection>
-        <LeftColumn>
-          <HeaderTitle>{spell.name} Buff Uptime</HeaderTitle>
-          <TimelineLabel>Timeline</TimelineLabel>
-        </LeftColumn>
-        <StatsColumn>
+    <SectionContainer>
+      <SectionHeader>
+        <TitleColumn>
+          <SectionTitle>{spell.name} Buff Uptime</SectionTitle>
+          <Label>Timeline</Label>
+        </TitleColumn>
+        <StatsRow>
           <StatCard color={backgroundBarColor}>
             <StatValue>{formatPercentage(uptimePercent, 0)}%</StatValue>
             <StatLabel>Uptime</StatLabel>
@@ -213,8 +225,8 @@ export default function BuffUptimeBar({
               <StatLabel>Max Stacks</StatLabel>
             </StatCard>
           )}
-        </StatsColumn>
-      </TopSection>
+        </StatsRow>
+      </SectionHeader>
 
       <TimelineContainer>
         <UptimeGraphContainer>
@@ -229,84 +241,12 @@ export default function BuffUptimeBar({
           />
         </UptimeGraphContainer>
       </TimelineContainer>
-    </Container>
+    </SectionContainer>
   );
 }
 
-const Container = styled.div`
-  margin-bottom: 16px;
-`;
-
-const TopSection = styled.div`
-  display: flex;
-  gap: 16px;
-  align-items: stretch;
-  margin-bottom: 8px;
-`;
-
-const LeftColumn = styled.div`
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  justify-content: space-between;
-`;
-
-const HeaderTitle = styled.h3`
-  margin: 0;
-  font-size: 18px;
-  font-weight: 600;
-  color: #fab700;
-  text-shadow: 0 2px 4px rgba(0, 0, 0, 0.5);
-`;
-
-const TimelineLabel = styled.div`
-  font-size: 11px;
-  font-weight: 600;
-  color: rgba(255, 255, 255, 0.6);
-  text-transform: uppercase;
-  letter-spacing: 1px;
-`;
-
-const StatsColumn = styled.div`
-  display: flex;
-  gap: 8px;
-  align-items: flex-end;
-`;
-
-const StatCard = styled.div<{ color: string }>`
-  background: rgba(0, 0, 0, 0.3);
-  border-radius: 6px;
-  padding: 6px 12px;
-  min-width: 70px;
-  border-left: 3px solid ${(props) => props.color};
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
-`;
-
-const StatValue = styled.div`
-  font-size: 18px;
-  font-weight: 700;
-  color: #fff;
-  line-height: 1;
-  margin-bottom: 2px;
-  text-shadow: 0 1px 3px rgba(0, 0, 0, 0.5);
-`;
-
-const StatLabel = styled.div`
-  font-size: 10px;
-  font-weight: 500;
-  color: rgba(255, 255, 255, 0.7);
-  text-transform: uppercase;
-  letter-spacing: 0.5px;
-`;
-
-const TimelineContainer = styled.div`
-  background: rgba(0, 0, 0, 0.3);
-  border-radius: 4px;
-  padding: 4px;
+const TimelineContainer = styled(InsetContainer)`
   height: 32px;
-  box-shadow: inset 0 1px 3px rgba(0, 0, 0, 0.4);
-  border: 1px solid rgba(255, 255, 255, 0.05);
 `;
 
 const UptimeGraphContainer = styled.div`
