+++ src/interface/guide/components/index.ts
@@ -1,6 +1,6 @@
 export { default as CastSummary, type CastEvaluation } from './CastSummary';
 export { default as GuideSection } from './GuideSection';
-export { default as StatisticCard, type StatisticData } from './StatisticCard';
+export { default as CastOverview, type StatisticData } from './CastOverview';
 export { default as CastDetail, type PerCastData, type PerCastStat } from './CastDetail';
 export { default as BuffUptimeBar } from './BuffUptimeBar';
 export {
@@ -9,3 +9,19 @@ export {
   type CastInSequence,
 } from './CastSequence';
 export { TipBox, TipBoxWithTimestamps } from './TipBox';
+export {
+  SectionContainer,
+  SectionHeader,
+  TitleColumn,
+  SectionTitle,
+  Label,
+  StatColumn,
+  StatsRow,
+  StatCard,
+  StatValue,
+  StatLabel,
+  StatsGrid,
+  StatItem,
+  StatItemValue,
+  StatItemLabel,
+} from './GuideDivs';
